fix(inference): resolve 8 pre-existing clippy warnings in bitnet-inference
- Add is_empty() to LruTracker (len_without_is_empty)
- Merge identical if-branches in detect_template() (if_same_then_else)
- Use div_ceil() in estimate_token_count() (manual_div_ceil)
- Use RangeInclusive::contains() in metric assertions (manual_range_contains)
- Remove unnecessary usize casts on argmax() (unnecessary_cast)
- Allow field_reassign_with_default in snapshot test (non_exhaustive struct)

+++ crates/bitnet-inference/tests/proptest_wave30_inference.rs
@@ -472,13 +472,13 @@ proptest! {
     #[test]
     fn argmax_within_bounds(logits in finite_f32_vec(1, 200)) {
         let idx = argmax(&logits);
-        prop_assert!((idx as usize) < logits.len());
+        prop_assert!(idx < logits.len());
     }
 
     /// argmax picks the actual maximum value.
     #[test]
     fn argmax_picks_max(logits in finite_f32_vec(1, 200)) {
-        let idx = argmax(&logits) as usize;
+        let idx = argmax(&logits);
         let max_val = logits.iter().cloned().fold(f32::NEG_INFINITY, f32::max);
         prop_assert!((logits[idx] - max_val).abs() < f32::EPSILON);
     }
